+++ NEWS.md
@@ -3,6 +3,8 @@
 * Content directories with a period in their name are no longer treated as a
   document path when computing the location for deployment records. (#1138)
 
+* A `_server.yml` file indicates that the content is an API. (#1144)
+
 # rsconnect 1.3.4
 
 * Use base64 encoded test data. Addresses CRAN test failures when run with